+++ AENetwork.xcodeproj/AENetwork.h
@@ -1,0 +1,4 @@
+#import <Foundation/Foundation.h>
+
+FOUNDATION_EXPORT double AENetworkVersionNumber;
+FOUNDATION_EXPORT const unsigned char AENetworkVersionString[];