+++ Network.xcodeproj/Network.h
@@ -1,4 +1,0 @@
-#import <Foundation/Foundation.h>
-
-FOUNDATION_EXPORT double NetworkVersionNumber;
-FOUNDATION_EXPORT const unsigned char NetworkVersionString[];